Updated the datatype in postgres and mysql for floats

+++ 1-setup/schema/university-mysql.sql
@@ -39,7 +39,7 @@ CREATE TABLE Student
     birth_date DATE,
     enrollment_date DATE,
     graduation_date DATE,
-    gpa FLOAT,
+    gpa FLOAT(4,2),
     UNIQUE (id)
 );
